requirements.txt
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.13.0
onnxruntime>=1.14.0
opencv-python>=4.7.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
Pillow>=9.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
streamlit>=1.22.0
grad-cam>=1.4.0
tqdm>=4.65.0
albumentations>=1.3.0
pydicom>=2.3.0
reportlab>=4.0.0
torchsummary>=1.5.1